misc fix

+++ include/graphic/display.h
@@ -26,13 +26,13 @@ struct plat_lcd_info {
 
 const struct lcd_vmode *lcd_get_vmode_by_id(int lcd_id);
 const struct lcd_vmode *lcd_get_vmode_by_name(const char *model);
-void *video_mem_alloc(__u32 *phy_addr, const struct lcd_vmode *vm, pixel_format_t pix_fmt);
+void *video_mem_alloc(unsigned long *phy_addr, const struct lcd_vmode *vm, pixel_format_t pix_fmt);
 
 //
 struct display {
 	void *mmio;
 	void *video_mem_va;
-	__u32   video_mem_pa;
+	__u32 video_mem_pa;
 
 	pixel_format_t pix_fmt;
 	struct lcd_vmode *video_mode;
